+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## v0.2.1 &rarr; master
 
-* iterator optimizaton (`nth` + `FusedIterator`)
+* iterator optimizatons (`nth` + `FusedIterator` + …)
 
 ## v0.1.0 &rarr; v0.2.1